backend/requirements.txt
fastapi==0.118.0
starlette==0.40.0
uvicorn[standard]==0.37.0
sqlalchemy==2.0.35
pymysql==1.1.2
alembic==1.16.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic-settings==2.11.0
python-dotenv==1.0.1
pydantic[email]==2.12.3